Refactor screen section so each screen has it's own stylesheet
