+++ package.json
@@ -12,11 +12,11 @@
     "react": "16.8.3",
     "react-dom": "16.8.3",
     "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
-    "react-native-gesture-handler": "^1.4.1",
     "react-native-normalize": "^1.0.1",
     "react-native-web": "^0.11.7",
     "react-navigation": "^4.0.10",
-    "react-navigation-stack": "^1.9.4"
+    "react-navigation-stack": "^1.9.4",
+    "react-native-gesture-handler": "~1.3.0"
   },
   "devDependencies": {
     "babel-preset-expo": "^7.0.0"
