+++ screens/AllGiantsScreen/AllGiantsStyles.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#48972C"
+  },
+  giantText: {
+    fontFamily: "amatic-sc",
+    fontSize: 60,
+    color: "white",
+    marginTop: "15%"
+  }
+});
+
+export default styles;
