+++ screens/MapScreen/MapScreen.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { View, Button, StyleSheet, Text, TouchableOpacity } from "react-native";
+import { View, Text, TouchableOpacity } from "react-native";
 
 export default class MapScreen extends React.Component {
   render() {
